refactor(engine): extract models::paths
Pure move: cache-root resolution, ModelKind, the is_cached/has_*_layout
probes and every FluidAudio location helper leave models/mod.rs for
models/paths.rs. mod.rs re-exports them, so no models::* call site moves.

The cache_dir tests split out of tts_tests with the code they exercise;
the four helpers mod.rs's remaining test modules reach for become
pub(super).

Refs #950
